# apis4lawenforcement_swaggerdef.yml
swagger: '2.0'
info:
  version: '1.0'
  title: "APIs 4 SAPOL Demo"
  description: SAPol provides a simple API allowing consumers to view people, license, vehicle and property info.
  #license:
  #  name: MIT
  #  url: https://github.com/apiaryio/polls-api/blob/master/LICENSE
...
basePath: "/"
host: "localhost:3000"
...
schemes:
- http
security:
  - X-MyAPI-Key: []
consumes:
- application/json
produces:
- application/json
paths:
  /tasks:
    x-summary: Tasks 
    post:
      tags: 
      - "Tasks"
      description: >-
        Create a Task representing a scope of work to be actioned. It may or may not be assigned to a Police Officer on createion. The assignment may be done later
      summary: Create a Task
      parameters:
        - name: body
          in: body
          required: true
          schema:
            properties:
              Task:
                $ref: '#/definitions/TaskRequest'
      responses:
        201:
          description: 'Successfully created Task Record'
          schema:
            properties:
              Task:
                $ref: '#/definitions/TaskResponse'
          examples:
            application/json:
              Task:
                taskId: 1
                taskName: 'Serve Summons to appear Ref ABC-1234 to Mr Craig Smith'
                taskDescription: 'Serve Summons to appear Ref ABC-1234 to Mr Craig Smith of 20 Charlesworthnut Street, Walkerville, SA 5000'
                taskLocation: '20 Charlesworthnut Street, Walkerville, SA 5000'
                taskStatus: 'In progress'
                taskStatusUpdateDate : '12-DEC-2018'
                lsa: 'Salisbury'
                officerAssignedToTask: '12345'
                officerAssignedDate: '12-DEC-2018'
                vixenBadgeId: '77777'

    get:
      tags: 
      - "Tasks"
      summary: "Get a List of All Tasks"
      description: "returns a list of tasks "
      operationId: "getTaskInfo"
      produces:
        - "application/json"
      parameters:
      - name: "badgeId"
        in: "query"
        description: "Enter a valid badgeId for an officer"
        required: false
        type: "string"
        x-example: 'A12345'
      - name: "lsa"
        in: "query"
        description: "Enter value of local service area eg Adelaide"
        required: false
        type: "string"
        x-example: 'Adelaide'
      responses:
        200:
          description: Successful Task Lookup Response
          schema:
            properties:
              Tasks:
                type: array
                items:
                  $ref: '#/definitions/TaskResponse'
          examples:
            application/json:
              Tasks:
                - taskId: 1
                  taskName: 'Serve Summons to appear Ref ABC-1234 to Mr Craig Smith'
                  taskDescription: 'Serve Summons to appear Ref ABC-1234 to Mr Craig Smith of 20 Charlesworthnut Street, Walkerville, SA 5000'
                  taskLocation: '20 Charlesworthnut Street, Walkerville, SA 5000'
                  taskStatus: 'In progress'
                  taskStatusUpdateDate : '12-DEC-2018'
                  las: 'Salisbury'
                  officerAssignedToTask: '12345'
                  officerAssignedDate: '12-DEC-2018'
                  vixenBadgeId: '77777'
                - taskId: 2
                  taskName: 'Interview Ms Sharon Smith'
                  taskDescription: 'Interview Ms Sharon Smith of 9 Ann Street, Salisbury, SA 5018 re stolen property - crime ref A3333'
                  taskLocation: '9 Ann Street, Salisbury, SA 5000'
                  taskStatus: 'Assigned'
                  taskStatusUpdateDate : '12-DEC-2018'
                  lsa: 'Salisbury'
                  officerAssignedToTask: '12345'
                  officerAssignedDate: '12-DEC-2018'
                  vixenBadgeId: '77777'
                - taskId: 3
                  taskName: 'Arrest Mike Downer'
                  taskDescription: 'Arrest Mr Mike Downer of 32 Owen Street, Salisbury Downs, SA 5019 re Serious Assualt - crime ref A44444'
                  taskLocation: '32 Owen Street, Salisbury Downs, SA 5109'
                  taskStatus: 'Created'
                  taskStatusUpdateDate : '12-JAN-2019'
                  lsa: 'Salisbury'
                  officerAssignedToTask: '' 
                  officerAssignedDate: ''
                  vixenBadgeId: '77777'
                - taskId: 4
                  taskName: 'Arrest Peter Pan'
                  taskDescription: 'Arrest Mr Peter Pan 25 Grenfell Street, Adelaide, SA 5000 re Serious Assualt - crime ref A44444'
                  taskLocation: '25 Grenfell Street, Adelaide, SA 5000'
                  taskStatus: 'Created'
                  taskStatusUpdateDate : '13-JAN-2019'
                  lsa: 'Adelaide'
                  officerAssignedToTask: '' 
                  officerAssignedDate: ''
                  vixenBadgeId: '88888'                
...
  /vehicles:
    x-summary: Vehicle Search 
    get:
      tags: 
      - "Vehicles"
      summary: List Vehicle Details 
      parameters:
      - name: "licensePlateNumber"
        in: "query"
        description: "the licenseplate of a vehicle"
        required: false
        type: "string"
        x-example: 'A333-ATV'
      - name: "registeredOwner"
        in: "query"
        description: "Enter the name of the registered owner"
        required: false
        type: "string"
        x-example: 'Mr Don Dundas'
      - name: "registeredOwnerAddress"
        in: "query"
        description: "Enter the address of the registered owner"
        required: false
        type: "string"
        x-example: '10 Thames Crescent, Salisbury East, SA, 5109'
      responses:
        200:
          description: Successful Vehicle Lookup Response
          schema:
            properties:
              Vehicles:
                type: array
                items:
                  $ref: '#/definitions/VehicleResponse'
          examples:
            application/json:
              Vehicles:
              - licensePlateNumber: 'A333-ATV'
                make: 'Mazda'
                model: 'CX-5'
                requiresLicenseClass: 'C'
                registrationStatus: 'Registered'
                registrationExpiryDate: '10-APR-2019'
                registeredOwner: 'Mr Don Dundas'
                registeredOwnerAddress: '10 Thames Crescent, Salisbury East, SA, 5109'
...
securityDefinitions:
  X-MyAPI-Key:
    type: "apiKey"
    name: "x-api-key"
    in: "header"
definitions:
  TaskRequest:
    title: "Task Request"
    type: "object"
    required:
    - "taskName"
    - "taskDescription"
    properties:
      taskName:
        type: string
        example: 'Serve Summons'
      taskDescription:
        type: string
        example: 'Detailed description of the task'
      officerAssignedToTask:
        type: string
        example: 'A12345'
      officerAssignedDate:
        type: string
        example: '20-JAN-2018'
      taskLocation:
        type: string
        example: 'Goodwood, SA'
      taskStatus:
        type: string
        example: 'Open'
      taskLastUpdatedDate:
        type: string
        example: '20-FEB-2018'
      lsa:
        type: string
        example: 'Adelaide'
      vixenBadgeId:
        type: string
        example: 'A77777'
  TaskResponse:
    title: "Task Response"
    type: object
    properties:
      taskId: 
        type: number
        example: 1
      taskName:
        type: string
        example: 'Serve Summons'
      taskDescription:
        type: string
        default: 'Detailed description of the task'
      officerAssignedToTask:
        type: string
        example: 'A12345'
      officerAssignedDate:
        type: string
        example: '20-JAN-2018'
      taskLocation:
        type: string
        example: 'Goodwood, SA'
      taskStatus:
        type: string
        example: 'Open'
      taskLastUpdatedDate:
        type: string
        example: '20-FEB-2018'
      lsa:
        type: string
        example: 'Adelaide'
      vixenBadgeId:
        type: string
        example: 'A77777'
...
  VehicleResponse:
    title: "Vehicle Response"
    type: object
    properties:
      licensePlateNumber:
        type: string
        example: 'S333-BUU'
      vehicleMake:
        type: string
        example: 'Mazda'
      vehicleModel:
        type: string
        example: 'CX-5'
      requiresLicenseClass:
        type: string
        example: 'C'
      registrationStatus: 
        type: string
        example: 'UNREGISTERED'
      registrationExpiryDate: 
        type: string
        example: '12-DEC-2018'
      registeredOwner:
        type: string
        example: 'Mr David Jones'
      registeredOwnerAddress:
        type: string
        example: '32 Angas Street, Adelaide, SA, 5000'
...